Basic setup for linearizing things
The main package (lh-array-sort) can be compiled in one of three modes now:

- mutable (primitive) arrays -- no checks, controlled by
  the (prim-)mutable-arrays cabal flags

- naive arrays, liquid checks, the default

- naive arrays, linear checks: pass `-f-liquid-checks` to `cabal build` to
  activate this

E.g. building in linear mode:

    > cabal build lh-array-sort -f-liquid-checks

To start linearizing things, build as shown above and
use -. as the linear arrow. Depending on the mode, -. resolves either
in the linear arrow or the usual arrow.

Syntactic caveat: functions with multiple arguments, some of which are
linear, must have explicit parenthesis; e.g. functions with two arguments:

    f :: a -. (b -> c)
    g :: a -> (b -. c)

(compare to `h :: a -> b -> c` which doesn't need the extra parenthesis)

The reason is that you can't match fixity of `->`, which is -1, according to
the GHC manual. User-defined operators can only have fixities 0 through 9.