Fix failing array tests (#11)
The test cases incorrectly assumed that makeArray initializes elements on
the mutable backends and that fromList is defined for size 0.
This commit updates the test suite to account for these issues by:

 - Removing tests that makeArray initializes elements for mutable backends.
 - Using replicate and fromList in splitMid tests to initialize elements.
 - Copying arrays when testing the commutativity of swaps.
 - Removing 0 from the range of fromList calls.

There is still one failing test. This is the test for swapping the same index,
which only fails on the prim-mutable-arrays backend.

Other changes:
 - Replace type annotations with type applications to improve readability.
 - Set the default-language of the test suite to GHC2021.